refactor namespace to AttributeResolver
