+++ tests/TestCase/AttributeResolver/ParserTest.php
@@ -14,11 +14,11 @@ declare(strict_types=1);
  * @since         6.0.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Attribute\Resolver;
+namespace Cake\Test\TestCase\AttributeResolver;
 
-use Cake\Attribute\Resolver\Enum\AttributeTargetType;
-use Cake\Attribute\Resolver\Parser;
-use Cake\Attribute\Resolver\ValueObject\AttributeInfo;
+use AttributeResolver\Enum\AttributeTargetType;
+use AttributeResolver\Parser;
+use AttributeResolver\ValueObject\AttributeInfo;
 use Cake\TestSuite\TestCase;
 use SplFileInfo;
 use TestApp\Attribute\Resolver\TestPriority;
